feat: withCallback & remove prefetch while csr
